Remove 'unresolved reference' issues, update bugs.
I was getting a lot of 'unresolved references' when generating for my
hyperlinks. Turns out to be a bug in Markdown -- if you end with a node
it doesn't expect you to end on (i.e. a list, or a blockquote) it'll
read the link references incorrectly.

I should fork it...