Big move to a clean build system.
Frog by default will do a lot at the top-level, and not be great about
cleaning up after itself either. I got around this before with
handcrafted scripts ('clean.sh') that did most of the job, but I never
liked that the toplevel had so much clutter.

Furthermore, if I wanted to 'productionize' the code (i.e. minify or
consolidate requests), I'd have to do it entirely around Frog.

This change marks the beginning of a new build process. On the surface,
we're just shuffling files and calling Frog from a new process, but we
can now grow things up a little, and the top-level is no longer
cluttered.

Since I don't know if my changes to Frog will make it upstream (and I'd
rather not wait until they do) I'm adding a private copy here.

Still much to be done!